Initial refactor complete
