+++ src/app/builder/edit/[buildId]/layout.tsx
@@ -1,6 +1,6 @@
 import BuildPage from './page'
 import { getBuild } from '../../actions'
-import { isErrorResponse } from '@/features/error-handling/lib/isErrorResponse'
+import { isErrorResponse } from '@/features/error-handling/isErrorResponse'
 
 export default async function Layout({
   params: { buildId },
