+++ src/app/builder/page.tsx
@@ -7,7 +7,7 @@ import useUrlBuildState from '@/features/build/hooks/useUrlBuildState'
 import { useIsClient } from 'usehooks-ts'
 import SaveBuildButton from '../../features/build/components/SaveBuildButton'
 import { ActionButton } from '../../features/build/components/ActionButton'
-import ToCsvButton from '../../features/csv/components/ToCsvButton'
+import ToCsvButton from '../../features/csv/ToCsvButton'
 import ImageDownloadLink from '../../features/build/components/ImageDownloadLink'
 import { useSession } from 'next-auth/react'
 import LoadingIndicator from '../../features/ui/LoadingIndicator'
