+++ src/app/profile/[userId]/layout.tsx
@@ -1,7 +1,7 @@
 import { Metadata, ResolvingMetadata } from 'next'
 import { getProfile } from './actions'
 import ProfilePage from './page'
-import { isErrorResponse } from '@/features/error-handling/lib/isErrorResponse'
+import { isErrorResponse } from '@/features/error-handling/isErrorResponse'
 
 export async function generateMetadata(
   { params: { userId } }: { params: { userId: string } },
