+++ src/app/profile/created-builds/page.tsx
@@ -2,17 +2,17 @@
 
 import BuildList from '@/features/build/components/BuildList'
 import BuildCard from '@/features/build/components/BuildCard'
-import usePagination from '@/features/pagination/hooks/usePagination'
+import usePagination from '@/features/pagination/usePagination'
 import { useEffect, useState } from 'react'
 import { CreatedBuildsFilter, getCreatedBuilds } from '../actions'
 import BuildListFilters from '@/features/build/components/BuildListFilters'
-import CopyBuildUrlButton from '../../../features/profile/components/CopyBuildUrlButton'
-import EditBuildButton from '../../../features/profile/components/EditBuildButton'
-import DuplicateBuildButton from '../../../features/profile/components/DuplicateBuildButton'
-import DeleteBuildButton from '../../../features/profile/components/DeleteBuildButton'
+import CopyBuildUrlButton from '../../../features/profile/CopyBuildUrlButton'
+import EditBuildButton from '../../../features/profile/EditBuildButton'
+import DuplicateBuildButton from '../../../features/profile/DuplicateBuildButton'
+import DeleteBuildButton from '../../../features/profile/DeleteBuildButton'
 import { DBBuild } from '@/features/build/types'
-import Tabs from '../../../features/profile/components/Tabs'
-import ProfileHeader from '../../../features/profile/components/ProfileHeader'
+import Tabs from '../../../features/profile/Tabs'
+import ProfileHeader from '../../../features/profile/ProfileHeader'
 import { useSession } from 'next-auth/react'
 import AuthWrapper from '@/features/auth/components/AuthWrapper'
