+++ src/app/profile/favorited-builds/page.tsx
@@ -1,16 +1,16 @@
 'use client'
 
-import CopyBuildUrlButton from '../../../features/profile/components/CopyBuildUrlButton'
+import CopyBuildUrlButton from '../../../features/profile/CopyBuildUrlButton'
 import { FavoritedBuildsFilter, getFavoritedBuilds } from '../actions'
-import usePagination from '@/features/pagination/hooks/usePagination'
+import usePagination from '@/features/pagination/usePagination'
 import { useEffect, useState } from 'react'
 import BuildCard from '@/features/build/components/BuildCard'
 import BuildListFilters from '@/features/build/components/BuildListFilters'
 import BuildList from '@/features/build/components/BuildList'
-import DuplicateBuildButton from '../../../features/profile/components/DuplicateBuildButton'
+import DuplicateBuildButton from '../../../features/profile/DuplicateBuildButton'
 import { DBBuild } from '@/features/build/types'
-import Tabs from '../../../features/profile/components/Tabs'
-import ProfileHeader from '../../../features/profile/components/ProfileHeader'
+import Tabs from '../../../features/profile/Tabs'
+import ProfileHeader from '../../../features/profile/ProfileHeader'
 import { useSession } from 'next-auth/react'
 import AuthWrapper from '@/features/auth/components/AuthWrapper'
